Update parallax sample to use fixed backgrounds
I think It didn't make sense for the backgrounds to be position:absolute - they'd scroll with the scroller and then be transformed more.  Even if wanted to show counter-transforms, in the slippage model this could look terrible (jitter).

Instead rely on the root document scrolling and make the backgrounds position:fixed.  Overflow scroller divs are also possible when the background isn't a descendent of the scroller, but that's more complex and less common.

@majido @flackr @ianvollick, agree?