[css-paint-api] Removes potential thread hopping from registerPaint and "draw a paint image".
This patch queues a task when registerPaint is called to update its map
of paint name to input properties.

It handles multiple calls by marking the name "invalid" if it recevies
different inputProperties with the same key.

"draw a paint image" is now split up into two functions.

1) "draw a paint image" which is the entry level function, performs the
checking if required, and
2) "invoke a paint callback" which is on the (potential) thread.

Fixes: #231, #232, #233.