Remove references to Animator object and message ports
The current spec does not have a browser-side Animator object, and it is no
longer possible to directly communicate from the browser to the Animation
worklet via message ports. In the future we will likely add a communication
method based on Element::getAnimations(), but for now get rid of the
inaccurate sections.