Remove all linking errors (#33)
The focus is on removing linking errors so that the output of
bikeshed.py is actually useful. Where possible I have tried to correct
the linking errors to what they should be, but if necessary I took the
side of removing erroring code when I couldn't work out how to fix it.