[css-animationworklet] Move threading model out of spec and into README (#904)
The Animation Worklet does not require or depend on the worklet running in a parallel execution context. In another word, it is valid and reasonable to run Animation Worklet on the main rendering thread. Similar to other Houdini primitives, the worklet is allowed to run on a different thread. This is an implementation detail that engine can take advantage of to improve animation smoothness.

Given that threading model is an implementation detail, I move most of that discussion to the Explainer.

The only aspect of that is relevant to the specification is "async" nature of worklet animation timing model (which enables the parallel execution mode). This is already discussed in "Worklet Animation timing model" and thus I moved the discussion of valid main thread sync opportunities for worklets to be part of that section.