[css-animationworklet-1] Add document definitions (#928)
Add document animator definition and corressponding map.

This ensures when registering an animator inside multiple global
scope, it is not valid to register different types of classes.
At the moment the only difference can be between statefullness.
This document animator map is also consulted when constructing
a worklet animation to reject use of an invalid (unregistered)
name.

Also updated the prose to use map from infra spec which is more
concrete and in line with paint worklet spec.