Editorial: Update use of WebIDL "invoke a callback function"
Rethrowing is no longer implicit (since reporting is frequently what's
desired). Here handling the exception is intended, so this is a very
small change.

Also included are the various changes required to make these specs build
and fixes to some places nearby where a clearly incorrect definition was
linked to.

Part of whatwg/webidl#1425.